requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
coverage==4.1
coveralls==1.1
decorator==4.0.9
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.5
django-allauth==0.25.2
django-debug-toolbar==1.4
django-envie==0.0.9
django-filter==0.13.0
django-registration==2.0.4
django-test==0.4030
djangorestframework==3.3.3
docopt==0.6.2
drfdocs==0.0.10
gnureadline==6.3.3
gunicorn==19.3.0
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
material-ui==0.1.3.6
nose==1.3.7
nose-pudb==1.0
oauthlib==1.1.2
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
pudb==2016.1
py==1.4.31
Pygments==2.1.3
PyJWT==1.4.0
python-openid==2.2.5
PyYAML==3.11
requests==2.10.0
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.19
static3==0.7.0
traitlets==4.2.1
urwid==1.3.1
wheel==0.24.0
